14.06.2023
-исправлен баг с переносом класса в domain(баг был из-за старой ссылки в nav graph)
-заменен kapt на ksp
-сдела небольшой рефакторинг пакетов

Tasks:

-перевести таймер на чистую архитектуру

-подумать над тем, чтобы сделать мульти таймеровое приложение, т.е. сделать возможность прекращать тренировку, а также пускать её на заднем фоне с кнопкой континью и минитаймером вместо start workout. Так можно будет вести несколько тренировок, но надо будет делать фабрики вместо синглов в дата слое и что-то придумать со звуками( в целом спорная фича)

-почистить код от мусора(старого кода, лишних картинок, стрингов и тд)

-сделать выбор при нажатии назад(отменить тренировку или нет) поменять кнопку старт на главной с старт на котинью
