+++ app/build.gradle
@@ -4,8 +4,9 @@ plugins {
     id 'kotlin-parcelize'
     // Navigation
     id 'androidx.navigation.safeargs.kotlin'
-    //Kotlin annotation processing tool for Room
-    id 'kotlin-kapt'
+    //Kotlin symbol processing  for Room
+    id("com.google.devtools.ksp")
+
 }
 
 android {
@@ -69,7 +70,8 @@ dependencies {
     // Room libraries
     implementation "androidx.room:room-runtime:$room_version"
     implementation "androidx.room:room-ktx:$room_version"
-    kapt "androidx.room:room-compiler:$room_version"
+    //Kotlin symbol  processing for Room
+    ksp "androidx.room:room-compiler:$room_version"
     //Koin
     implementation "io.insert-koin:koin-core:$koin_version"
     implementation "io.insert-koin:koin-android:$koin_version"
