+++ app/src/main/java/com/practice/getup/domain/models/Workout.kt
@@ -1,0 +1,14 @@
+package com.practice.getup.domain.models
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class Workout(
+    val id: Int = 0,
+    val name: String,
+    val preparingTime: Int,
+    val workTime: Int,
+    val restTime: Int,
+    val numberOfSets: Int
+): Parcelable
