+++ app/src/main/java/com/practice/getup/domain/storage/DeleteWorkoutUseCase.kt
@@ -1,6 +1,6 @@
 package com.practice.getup.domain.storage
 
-import com.practice.getup.database.Workout
+import com.practice.getup.domain.models.Workout
 import com.practice.getup.domain.repositories.StorageRepository
 
 class DeleteWorkoutUseCase(private val storageRepository: StorageRepository) {
