+++ app/src/main/java/com/practice/getup/domain/storage/GetWorkoutUseCase.kt
@@ -1,8 +1,7 @@
 package com.practice.getup.domain.storage
 
-import com.practice.getup.database.Workout
+import com.practice.getup.domain.models.Workout
 import com.practice.getup.domain.repositories.StorageRepository
-import kotlinx.coroutines.flow.Flow
 
 class GetWorkoutUseCase(private val storageRepository: StorageRepository) {
