+++ app/src/main/java/com/practice/getup/presentation/main_menu/view_model/MainMenuViewModel.kt
@@ -4,13 +4,9 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
-import com.practice.getup.database.Workout
-import com.practice.getup.domain.storage.AddNewWorkoutUseCase
+import com.practice.getup.domain.models.Workout
 import com.practice.getup.domain.storage.DeleteAllWorkoutsUseCase
-import com.practice.getup.domain.storage.DeleteWorkoutUseCase
 import com.practice.getup.domain.storage.GetAllWorkoutsUseCase
-import com.practice.getup.domain.storage.GetWorkoutUseCase
-import com.practice.getup.domain.storage.UpdateWorkoutUseCase
 import kotlinx.coroutines.launch
 
 class MainMenuViewModel(
