+++ build.gradle
@@ -20,5 +20,5 @@ plugins {
     id 'com.android.application' version '7.4.0' apply false
     id 'com.android.library' version '7.4.0' apply false
     id 'org.jetbrains.kotlin.android' version '1.8.21' apply false
-
+    id("com.google.devtools.ksp") version "1.8.22-1.0.11" apply false
 }
